+++ src/index.test.tsx
@@ -5,6 +5,11 @@ describe("index", () => {
     const root = document.createElement("div");
     root.setAttribute("id", "root");
     document.body.appendChild(root);
+
+    const themeToggle = document.createElement("theme-toggle");
+    themeToggle.setAttribute("id", "theme-toggle");
+    document.body.appendChild(themeToggle);
+
     require("./index");
     // @ts-ignore
     expect(document.getElementById("root")._reactRootContainer).toBeTruthy();